+++ examples/sqlite/src/query_root.rs
@@ -3,6 +3,8 @@ use async_graphql::dynamic::*;
 use sea_orm::DatabaseConnection;
 use seaography::{async_graphql, lazy_static, Builder, BuilderContext};
 
+mod mutations;
+
 lazy_static::lazy_static! { static ref CONTEXT : BuilderContext = BuilderContext :: default () ; }
 
 pub fn schema(
@@ -33,9 +35,7 @@ pub fn schema(
         ]
     );
 
-    builder
-        .mutations
-        .extend([Foo::gql(), Bar::gql(), Login::gql()]);
+    builder.mutations.extend(mutations::def());
 
     builder
         .set_depth_limit(depth)
@@ -43,109 +43,4 @@ pub fn schema(
         .schema_builder()
         .data(database)
         .finish()
-}
-
-use async_graphql::Result as GqlResult;
-use seaography::macros::CustomMutation;
-
-/*
-fn foo() -> gql_dyn::Field {
-    gql_dyn::Field::new(
-        "foo",
-        <String as seaography::AsyncGqlValueType>::gql_type_ref(&CONTEXT),
-        move |ctx| {
-            FieldFuture::new(async move {
-                let username = <String as seaography::AsyncGqlValueType>::try_get_arg(
-                    &CONTEXT, &ctx, "username",
-                )?;
-
-                let result = format!("Hello, {}!", username);
-                Ok(Some(gql_dyn::FieldValue::value(result)))
-            })
-        },
-    )
-    .argument(gql_dyn::InputValue::new(
-        "username",
-        <String as seaography::AsyncGqlValueType>::gql_type_ref(&CONTEXT),
-    ))
-}
-*/
-#[derive(CustomMutation)]
-struct Foo {
-    foo: fn(username: String) -> String,
-}
-
-impl Foo {
-    async fn foo(_ctx: &ResolverContext<'_>, username: String) -> GqlResult<String> {
-        Ok(format!("Hello, {}!", username))
-    }
-}
-
-/*
-fn bar() -> gql_dyn::Field {
-    gql_dyn::Field::new(
-        "bar",
-        <i32 as seaography::AsyncGqlValueType>::gql_type_ref(&CONTEXT),
-        move |ctx| {
-            FieldFuture::new(async move {
-                let x = <i32 as seaography::AsyncGqlValueType>::try_get_arg(&CONTEXT, &ctx, "x")?;
-                let y = <i32 as seaography::AsyncGqlValueType>::try_get_arg(&CONTEXT, &ctx, "y")?;
-
-                let result = x + y;
-                Ok(Some(gql_dyn::FieldValue::value(result)))
-            })
-        },
-    )
-    .argument(gql_dyn::InputValue::new(
-        "x",
-        <i32 as seaography::AsyncGqlValueType>::gql_type_ref(&CONTEXT),
-    ))
-    .argument(gql_dyn::InputValue::new(
-        "y",
-        <i32 as seaography::AsyncGqlValueType>::gql_type_ref(&CONTEXT),
-    ))
-}
-*/
-#[derive(CustomMutation)]
-struct Bar {
-    bar: fn(x: i32, y: i32) -> i32,
-}
-
-impl Bar {
-    async fn bar(_ctx: &ResolverContext<'_>, x: i32, y: i32) -> GqlResult<i32> {
-        Ok(x + y)
-    }
-}
-
-/*
-fn login() -> gql_dyn::Field {
-    gql_dyn::Field::new(
-        "login",
-        <seaography::SeaOrmModel<customer::Model> as seaography::AsyncGqlValueType>::gql_type_ref(
-            &CONTEXT,
-        ),
-        move |ctx| {
-            FieldFuture::new(async move {
-                use sea_orm::EntityTrait;
-
-                let repo = ctx.data::<DatabaseConnection>().unwrap();
-
-                let result = customer::Entity::find().one(repo).await?.unwrap();
-                Ok(Some(gql_dyn::FieldValue::owned_any(result)))
-            })
-        },
-    )
-}
-*/
-#[derive(CustomMutation)]
-struct Login {
-    login: fn() -> customer::Model,
-}
-impl Login {
-    async fn login(ctx: &ResolverContext<'_>) -> GqlResult<customer::Model> {
-        use sea_orm::EntityTrait;
-
-        let repo = ctx.data::<DatabaseConnection>().unwrap();
-        Ok(customer::Entity::find().one(repo).await?.unwrap())
-    }
 }
